requirements.txt
pyftpdlib>=1.5.9
watchdog>=4.0
pyyaml>=6.0
rumps>=0.4.0
Pillow>=10.0
requests>=2.28

